requirements.txt
Django==2.1.4
djangorestframework==3.9.0
gunicorn==19.9.0
joblib==0.13.0
numpy==1.15.4
psycopg2==2.7.6.1
pytz==2018.7
scikit-learn==0.20.1
scipy==1.1.0
sklearn==0.0
